+++ src/components/app/app.tsx
@@ -16,8 +16,8 @@ import ResetPassword from "../../pages/reset-password";
 import Profile from "../../pages/profile";
 import ProfileMenu from "../../pages/profile-menu";
 import ProfileOrders from "../../pages/profile-orders";
-import IngredientDetails from "../burger-ingredients/ingredient-details";
-import IngredientModal from "../burger-ingredients/ingredient-modal";
+import IngredientDetails from "../ingredient-details/ingredient-details";
+import IngredientModal from "../ingredient-details/ingredient-modal";
 import { AuthenticatedOnly, AnonymousOnly } from "../protected-route";
 import OrderDescription from "../order-description/order-description";
 import OrderDescriptionModal from "../order-description/order-description-modal";
