+++ src/components/ingredient-details/ingredient-details.tsx
